refactor: modularizes ts types and functions
